push the github...

+++ command/owner/say.js
@@ -1,4 +1,19 @@
-exports.execute = async function (msg, command, args, client, D, perm, color) {
-msg.delete()
-msg.channel.send(args.join(" "))
+exports.execute = function (msg, command, args, client, D, perm, color) {
+    if(!args) return;
+
+    let newargs = []
+    let delthis = false
+    args.forEach(str => {
+        if (str === "-d" || str == "-del") {
+            delthis = true
+            return
+        }
+        newargs.push(str)
+    });
+    if(delthis == true){
+        msg.delete()
+        msg.channel.send(`${newargs.join(" ")}`)
+    }else{
+        msg.channel.send(`${newargs.join(" ")}`)
+    }
 }
